+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Changed
 * Changed not to use memcached.
   Contributed by @hinashi
+* Changed not to execute notify job when there is no update in entry import
+  Contributed by @hinashi
 
 ### Fixed